IEGO Template exporter was not functioning in game
The IEGO template could no longer be worked ingame because the add-on detected the IEGO template (to use the XRES format) based on its name, but the template had been renamed (Inazuma Eleven Go to IEGO). A file_version value has been added to the template class and json. When it is set to 1, the add-on understands that it must use xres and v1 animation. When it is set to 2 or undefined, the add-on will use classic res and v2 animation.